Made yaml form general and adapted decoders
