+++ public/pages/Decoders/components/DecoderDetailsFlyout.tsx
@@ -21,15 +21,14 @@ import {
   EuiSpacer,
   EuiText,
 } from '@elastic/eui';
-import { DecoderItem } from '../../../../types';
+import { DecoderDocument, DecoderItem } from '../../../../types';
 import { DataStore } from '../../../store/DataStore';
 import { EnabledHealth } from '../../../components/Utility/EnabledHealth';
 import { Metadata, MetadataFieldType } from '../../../components/Utility/Metadata';
 import { DEFAULT_EMPTY_DATA } from '../../../utils/constants';
 import { BadgeGroup } from '../../../components/Utility/BadgeGroup';
 import { stringify as LosslessStringify } from 'lossless-json';
-import { mapYamlToLosslessDecoder } from '../components/mappers';
-
+import { mapYamlToLosslessObject } from '../../../components/YamlForm';
 interface DecoderDetailsFlyoutProps {
   decoderId: string;
   space: string;
@@ -99,7 +98,7 @@ export const DecoderDetailsFlyout: React.FC<DecoderDetailsFlyoutProps> = ({
     try {
       const rawYaml = typeof decoder.yaml === 'string' ? decoder.yaml : null;
       if (rawYaml) {
-        const losslessDoc = mapYamlToLosslessDecoder(rawYaml);
+        const losslessDoc = mapYamlToLosslessObject<DecoderDocument>(rawYaml);
         return LosslessStringify(losslessDoc, null, 2) ?? '';
       }
       return JSON.stringify(decoder?.document, null, 2);
@@ -131,7 +130,7 @@ export const DecoderDetailsFlyout: React.FC<DecoderDetailsFlyoutProps> = ({
           values={decoder?.document?.metadata?.supports}
         />
       ),
-      type: 'raw'
+      type: 'raw',
     },
   ];
