+++ public/pages/Decoders/utils/constants.ts
@@ -9,6 +9,16 @@ const TEXT_SEARCH_FIELDS = ['document.metadata.title', 'document.metadata.descri
 
 const escapeWildcard = (str: string) => str.replace(/[*?]/g, '\\$&');
 
+export const decoderFormDefaultValue: string = `name: decoder/<name>/<version>
+enabled: true
+metadata:
+  title: Placeholder Decoder
+  description: This is a placeholder decoder. Please update the fields accordingly.
+  author: User
+  references: []
+  documentation: ''
+  supports: []`;
+
 export const buildDecodersSearchQuery = (searchText: string) => {
   const trimmed = searchText.trim();
   if (!trimmed) {
